# docs/openapi/mintmory.yaml
openapi: 3.1.0
info:
  title: MintMory Memory API
  version: 0.1.0
  description: |
    HTTP API for MintMory, a typed memory system for LLM agents.

    ## Design Model

    MintMory follows a **Palantir Ontology-inspired** data model:

    - **Object Types** → `MemoryRecord` (primary memory instances), `ConceptLink`
      (typed edges between memories), `QuerySession` (decision-lineage audit records),
      `MemorySummary` (LLM-synthesised concept digests).
    - **Link Types** → 12 orthogonal `ConceptLinkType` values drawn from the memory
      graph literature (KGFoller, AgentKB, Recall-and-Reflect, StructGPT, Memento,
      MemFree). Includes `annotates` (asymmetric: note → item).
    - **Action Types** → exposed both as MCP tools (primary transport) and as the REST
      endpoints documented here.
    - **Decision Lineage** → every search creates a `QuerySession`; the LLM self-reports
      which memories were useful or stale via `POST /sessions/{id}/feedback`.

    ## Transports

    This REST API and the MCP tool layer are **parallel transports over the same service
    layer** — every endpoint maps 1-to-1 to an MCP tool. Prefer the MCP transport when
    operating inside an LLM agent context; use REST for tooling, dashboards, and
    service-to-service integration.

    ## Scoring Model

    Memory retrieval rank uses a 7-factor composite score (arxiv 2606.12945):

    ```
    effective_score = base_relevance
                    + usefulness_score × 0.3
...
    ```

...
servers:
...
paths:
...
  /memories:
    post:
      operationId: addMemory
      summary: Add a new memory
      description: |
        Stores a new `MemoryRecord`. The service will:
        1. Generate a ULID-format `id` if none is provided.
        2. Run entity extraction to populate `entity_ids`.
        3. Run contradiction detection against recent active memories; sets
           `flagged_for_review` and `contradicts_ids` if conflicts are found.
        4. Infer initial `ConceptLink` edges from the extracted entities.

        Returns `409 Conflict` if a memory with semantically identical content
        already exists (cosine similarity ≥ 0.98) and deduplication is enabled.
      tags:
        - Memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryCreate'
            examples:
              basicFact:
                summary: Add a simple fact memory
                value:
                  content: "The Acme parking integration uses OAuth 2.0 with PKCE and
                    requires the `parking:read` scope."
                  category: fact
                  source: agent
                  confidence: 0.95
                  metadata:
                    project_id: "acme-parking"
                    session_id: "ses_01HXYZ"
      responses:
        '201':
          description: Memory created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryRecord'
              examples:
                created:
                  summary: Created memory record
                  value:
                    id: "01HXYZ1234567890ABCDEFGHIJ"
                    content: "The Acme parking integration uses OAuth 2.0 with PKCE and
                      requires the `parking:read` scope."
                    category: fact
                    entity_ids: ["Acme", "OAuth 2.0", "PKCE", "parking:read"]
                    confidence: 0.95
                    source: agent
                    verified: false
                    valid_from: null
                    valid_until: null
                    usefulness_score: 0.0
                    staleness_score: 0.0
                    reinforcement_count: 0
                    retrieval_count: 0
                    is_stale: false
                    is_active: true
                    is_archived: false
                    superseded_by: null
                    flagged_for_review: false
                    contradicts_ids: []
                    created_at: "2026-06-15T10:00:00Z"
                    updated_at: "2026-06-15T10:00:00Z"
                    last_accessed_at: null
                    metadata:
                      project_id: "acme-parking"
                      session_id: "ses_01HXYZ"
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /memories/{id}:
    parameters:
      - $ref: '#/components/parameters/MemoryId'

    get:
      operationId: getMemory
      summary: Get memory by ID
      description: |
        Retrieves a single `MemoryRecord` by its ID. Increments `retrieval_count`
        and updates `last_accessed_at` as a side-effect.
      tags:
        - Memories
      responses:
        '200':
          description: The requested memory record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryRecord'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateMemory
      summary: Update memory fields
      description: |
        Partially updates a `MemoryRecord`. Only the fields present in the request
        body are changed. Updatable fields: `content`, `category`, `verified`,
        `metadata`, `valid_from`, `valid_until`, `confidence`.

        Setting `content` triggers a fresh entity extraction and contradiction
        check. `updated_at` is always refreshed.
      tags:
        - Memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryUpdate'
      responses:
        '200':
          description: Updated memory record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryRecord'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      operationId: archiveMemory
      summary: Archive a memory (soft delete)
      description: |
        Soft-deletes a `MemoryRecord` by setting `is_archived = true` and
        `is_active = false`. The record is retained in storage for audit and
        lineage purposes but excluded from searches by default.

        To permanently remove a record, use the admin CLI — the HTTP API does
        not expose hard-delete.
      tags:
        - Memories
      responses:
        '204':
          description: Memory archived. No response body.
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Search
  # ---------------------------------------------------------------------------
  /memories/search:
    post:
      operationId: searchMemories
      summary: Search memories
      description: |
        Performs semantic vector search over active memories, optionally followed
        by graph traversal (search-around).

        **Phase 1 — vector search**: embeds `query` and retrieves the top-`limit`
        memories ranked by `effective_score` (composite: relevance + usefulness
        - staleness + recency bonus).

        **Phase 2 — search-around** (optional): traverses `ConceptLink` edges from
        the Phase 1 anchor set using the `search_around` spec (link types, depth,
        min_strength). Retrieved neighbours are unioned with the anchor set.
        `search_around_ids` in the response identifies the graph-traversal additions.

        Creates a `QuerySession` record; the returned `session_id` must be used in
        `POST /sessions/{id}/feedback` to close the feedback loop.
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              basicSearch:
                summary: Simple keyword search
                value:
                  query: "OAuth integration parking"
                  limit: 5
              searchAroundWithFilter:
                summary: Search with graph traversal and filter
                value:
                  query: "parking integration"
                  search_around:
                    link_types: ["contradicts", "supersedes"]
                    depth: 2
                    min_strength: 0.4
                  filter:
                    category: fact
                    active_only: true
                    exclude_stale: true
                    min_usefulness: 1.5
                  limit: 10
                  include_archived: false
      responses:
        '200':
          description: Search results with session tracking ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              examples:
                results:
                  summary: Successful search response
                  value:
                    session_id: "ses_01HXYZ9999"
                    memories:
                      - id: "01HXYZ1234567890ABCDEFGHIJ"
                        content: "The Acme parking integration uses OAuth 2.0 with PKCE."
                        category: fact
                        entity_ids: ["Acme", "OAuth 2.0", "PKCE"]
                        confidence: 0.95
                        source: agent
                        verified: false
                        valid_from: null
                        valid_until: null
                        usefulness_score: 3.2
                        staleness_score: 0.1
                        reinforcement_count: 4
                        retrieval_count: 12
                        is_stale: false
                        is_active: true
                        is_archived: false
                        superseded_by: null
                        flagged_for_review: false
                        contradicts_ids: []
                        created_at: "2026-06-10T08:00:00Z"
                        updated_at: "2026-06-14T09:15:00Z"
                        last_accessed_at: "2026-06-15T09:55:00Z"
                        metadata:
                          project_id: "acme-parking"
                    total_found: 1
                    search_around_ids: []
        '400':
          $ref: '#/components/responses/ValidationError'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ---------------------------------------------------------------------------
  # Links
  # ---------------------------------------------------------------------------
  /memories/{id}/links:
    parameters:
      - $ref: '#/components/parameters/MemoryId'

    get:
      operationId: getMemoryLinks
      summary: Get all concept links for a memory
      description: |
        Returns every `ConceptLink` where this memory appears as either
        `source_memory_id` or `target_memory_id`. The response is a flat list
        of link objects; inspect `source_memory_id` and `target_memory_id` to
        determine direction.

        To filter by link type, use the `link_type` query parameter.
      tags:
        - Links
      parameters:
        - name: link_type
          in: query
          required: false
          description: |
            Filter results to only links of this type. When omitted, all 12
            `ConceptLinkType` values are included.
          schema:
            $ref: '#/components/schemas/ConceptLinkType'
        - name: min_strength
          in: query
          required: false
          description: |
            Exclude links with `strength` below this threshold (0–1).
            Defaults to 0 (all links returned).
          schema:
            type: number
            format: float
            minimum: 0.0
            maximum: 1.0
            default: 0.0
      responses:
        '200':
          description: List of concept links involving this memory.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConceptLink'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      operationId: createMemoryLink
      summary: Create a manual concept link
      description: |
        Creates a new `ConceptLink` from this memory (as `source_memory_id`) to
        another memory. The `link_type` determines the semantic relationship.

        `source` is always set to `user` for links created via this endpoint.

        For symmetric link types (`relates_to`, `contradicts`, `similar_to`) the
        server will also create the reverse link automatically.
      tags:
        - Links
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConceptLinkCreate'
      responses:
        '201':
          description: Concept link created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptLink'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ---------------------------------------------------------------------------
  # Sessions
  # ---------------------------------------------------------------------------
  /sessions/{id}:
    parameters:
      - $ref: '#/components/parameters/SessionId'

    get:
      operationId: getSession
      summary: Get a query session
      description: |
        Retrieves a `QuerySession` record by its ID. Includes all post-session
        feedback fields if feedback has already been submitted (`concluded: true`).
      tags:
        - Sessions
      responses:
        '200':
          description: The requested query session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuerySession'
        '404':
          $ref: '#/components/responses/NotFound'
  /sessions/{id}/feedback:
    parameters:
      - $ref: '#/components/parameters/SessionId'

    post:
      operationId: submitSessionFeedback
      summary: Submit LLM-assessed session feedback
      description: |
        Records the LLM's self-assessment of which memories were useful and which
        were stale, closing the feedback loop for a `QuerySession`.

        This endpoint implements **passive self-assessment** (Recall-and-Reflect
        pattern): the LLM calls it automatically after answering a user query —
        no human interaction is required.

        Side-effects:
        - `concluded` on the session is set to `true`.
        - `usefulness_score` of memories in `useful_ids` is incremented.
        - `staleness_score` of memories in `stale_ids` is incremented; memories
          may be automatically flagged as `is_stale` if threshold is crossed.
        - `reinforcement_count` on useful memories is incremented.

        Returns `409 Conflict` if feedback was already submitted for this session.
      tags:
        - Sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionFeedback'
      responses:
        '200':
          description: Updated query session with feedback applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuerySession'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  # ---------------------------------------------------------------------------
  # Dream
  # ---------------------------------------------------------------------------
...
  /summaries:
    get:
      operationId: listSummaries
      summary: List all memory summaries
      description: |
        Returns all `MemorySummary` records ordered by `generated_at` descending.
        Summaries are created and updated by the dreaming process; this endpoint
        is read-only.

        Only `is_current = true` summaries are returned by default. Set
        `include_stale=true` to include summaries whose linked memories have
        since become stale.
      tags:
        - Summaries
      parameters:
        - name: include_stale
          in: query
          required: false
          description: |
            When `true`, summaries with `is_current = false` are included in the
            response. Defaults to `false`.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of memory summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemorySummary'
...
components:
  parameters:
    MemoryId:
      name: id
      in: path
      required: true
      description: |
        Unique identifier of the `MemoryRecord`. ULID format (26 chars,
        base32 Crockford), e.g. `01HXYZ1234567890ABCDEFGHIJ`.
      schema:
        type: string
        minLength: 1
...
  schemas:
...
    MemoryCreate:
      type: object
      description: |
        Request body for adding a new memory (`POST /memories`). Only `content`
        and `category` are required; all other fields have server-side defaults.
      required:
        - content
        - category
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000
          description: The natural-language content of the memory to store.
          example: "The Acme parking integration uses OAuth 2.0 with PKCE."
        category:
          $ref: '#/components/schemas/MemoryCategory'
        source:
          $ref: '#/components/schemas/MemorySource'
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          default: 1.0
          description: Agent's confidence that this memory is accurate.
          example: 0.95
        verified:
          type: boolean
          default: false
          description: Whether this memory has already been verified by the user.
          example: false
        valid_from:
          type: string
          format: date-time
          nullable: true
          description: Start of validity window for temporal memories.
          example: null
        valid_until:
          type: string
          format: date-time
          nullable: true
          description: End of validity window for temporal memories.
          example: null
        metadata:
          type: object
          additionalProperties: true
          default: {}
          description: Arbitrary application-level metadata.
          example:
            project_id: "acme-parking"
...